requirements.txt
influxdb>=5.3.1
requests>=2.31.0
python-dateutil>=2.8.2
pyyaml>=6.0.1
pytz>=2024.1
tqdm>=4.66.0 